refactor: :art: Separation of map concerns
