+++ src/scripts/map.ts
@@ -1,8 +1,5 @@
 import { map as createMap, latLng, marker, tileLayer, Map } from "leaflet";
-import { weatherFetchRequest$ } from "./api.ts";
-
 import { ICityData } from "../models/cityData";
-import { mapDisplayToggle } from "./dom.ts";
 
 export function updateMap(map: Map, cityData: ICityData): void {
   map.panTo(latLng(cityData.latitude, cityData.longitude));
@@ -28,30 +25,4 @@ export function initilizeMap(currentCity: ICityData): Map {
     .addTo(myMap)
     .openPopup();
   return myMap;
-}
-
-export function mapListeners(
-  map: Map,
-  mapDiv: HTMLElement,
-  sevenDayDiv: HTMLElement,
-  openMap: HTMLElement
-): void {
-  openMap.addEventListener("click", () =>
-    mapDisplayToggle(mapDiv, sevenDayDiv)
-  );
-
-  map.on("click", onMapClick);
-
-  function onMapClick(e: { latlng: { lat: number; lng: number } }) {
-    let longitude = Math.round(e.latlng.lat * 1000) / 1000;
-    let latitude = Math.round(e.latlng.lng * 1000) / 1000;
-    const cityData: ICityData = {
-      longitude: longitude,
-      latitude: latitude,
-      cityName: "Selected Location",
-    };
-    weatherFetchRequest$.next(cityData);
-    mapDiv.style.display = "none";
-    sevenDayDiv.style.display = "flex";
-  }
 }
